// planner-web/src/api/generated/model/financialPlanDetailResponseDto.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Financial Planner API
 * Versioned API for normalized financial plans
 * OpenAPI spec version: 1.0
 */
import type { AccountResponseDto } from "./accountResponseDto"
import type { AllocationCategoryResponseDto } from "./allocationCategoryResponseDto"
import type { ChildrenBufferRuleDto } from "./childrenBufferRuleDto"
import type { CompletedItemResponseDto } from "./completedItemResponseDto"
import type { CreditCardBackingRuleDto } from "./creditCardBackingRuleDto"
import type { DebtLiquidationRuleDto } from "./debtLiquidationRuleDto"
import type { FinancialPlanDetailResponseDtoStatus } from "./financialPlanDetailResponseDtoStatus"
import type { IncomePaymentResponseDto } from "./incomePaymentResponseDto"
import type { IncomeScheduleResponseDto } from "./incomeScheduleResponseDto"
import type { IncomeScheduleRuleDto } from "./incomeScheduleRuleDto"
import type { PaymentPeriodResponseDto } from "./paymentPeriodResponseDto"
import type { RecurringExpenseResponseDto } from "./recurringExpenseResponseDto"
import type { WeeklyFlexibleBudgetRuleDto } from "./weeklyFlexibleBudgetRuleDto"

export interface FinancialPlanDetailResponseDto {
  id: string
  metadataId: string
  schemaVersion: string
  name: string
  currency: string
  startDate: string
  /** @nullable */
  endDate: string | null
  status: FinancialPlanDetailResponseDtoStatus
  /** @nullable */
  objective: string | null
  /** @nullable */
  projectedDebtFreeDate: string | null
  /** @nullable */
  projectedEmergencyFund: number | null
  createdAt: string
  updatedAt: string
  allocationCategories: AllocationCategoryResponseDto[]
  accounts: AccountResponseDto[]
  /** @nullable */
  incomeSchedule: IncomeScheduleResponseDto | null
  incomePayments: IncomePaymentResponseDto[]
  paymentPeriods: PaymentPeriodResponseDto[]
  recurringExpenses: RecurringExpenseResponseDto[]
  completedItems: CompletedItemResponseDto[]
  rules: (
    | IncomeScheduleRuleDto
    | ChildrenBufferRuleDto
    | WeeklyFlexibleBudgetRuleDto
    | CreditCardBackingRuleDto
    | DebtLiquidationRuleDto
  )[]
}